faker/requirements.txt
Faker==15.3.3
mysql_connector_repackaged==0.3.1
